parte-5

+++ src/main/java/com/sala/facil/DTOS/ReservaDTO.java
@@ -5,6 +5,6 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDateTime;
 
-public record ReservaDTO(long id_reserva, @NotNull LocalDateTime data_reserva, @NotNull LocalDateTime data_pedido,
-                         @NotNull int status, @NotNull int sala_id, @NotNull int usuario_id) {
+public record ReservaDTO(Long id_reserva, @NotNull LocalDateTime data_reserva, @NotNull LocalDateTime data_pedido,
+                         @NotNull int status, @NotNull Long sala_id, @NotNull Long usuario_id) {
 }
